Harden the parsers against crafted volumes and cache files
A fuzzing review of everything that reads untrusted bytes, roughly
twelve million iterations under the sanitizers. It found no
out-of-bounds access anywhere, which is the part worth stating plainly,
and six integer, allocation and semantic bugs, which are these.

A crafted cache file could have taken the process down. The reader
reserves each node's children up front, and the node count it trusts
was bounded only by the file size, so a one gigabyte cache could demand
close to four gigabytes of address space in a single call. The count is
now bounded by a memory budget instead, and the two calls that load a
cache on the interface thread are wrapped, because an allocation
failure escaping a window procedure is not an error message.

A crafted volume could have taken the scan down. The five largest
allocations in the MFT reader sat outside the guard that protects the
two smaller ones, so instead of the documented fall back to the
directory walk, a failure ended the scan, and in the headless export
ended the process. They are inside it now, and a failure leaves no
half-built tree behind.

A crafted volume could also have lied about where a file lives. The
parent reference in a file name record is forty-eight bits and was
narrowed to thirty-two before being range checked, so an index far past
the table wrapped onto a valid one, most usefully onto the root: the
file was grafted to the volume root, and every path shown for it, or
acted on, named somewhere it did not live. The reference is now kept
and compared at full width. Three multiplications in the read plan
saturate rather than wrapping, so a hostile run list can no longer
produce a plausible but wrongly attributed tree.

Two smaller ones: the anti-replay serial in the settings file, added
earlier today, accumulated digits without an overflow check, so a value
above the maximum wrapped to zero and read as "accept anything"; and a
size filter with a long fraction wrapped to a single byte, which turned
"larger than a terabyte" into "larger than nothing". The fraction is
now bounded where it must be, which is six digits given the terabyte
multiplier, and still resolves a megabyte in a terabyte.

Also aligned the depth limits: the walker built one level deeper than
either reader would accept, so a legitimately deep tree's cache was
written and then rejected and deleted on every launch.

948 host assertions pass under the sanitizers, analysis is clean.

+++ src/ntfs.cpp
@@ -351,8 +351,8 @@ RecordInfo ParseRecord(const uint8_t* record, size_t len) {
                         }
 
                         bestNamespace = rank;
-                        info.parent = static_cast<uint32_t>(
-                            r.U64(base) & 0x0000FFFFFFFFFFFFull);
+                        info.parent =
+                            r.U64(base) & 0x0000FFFFFFFFFFFFull;
 
                         info.name.clear();
                         info.name.reserve(nameChars);
